backend/requirements.txt
# Django Core
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==23.5

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# Database (if needed)
# psycopg2-binary==2.9.9  # For PostgreSQL

# Development Tools
python-decouple==3.8

# Additional Utilities
Pillow==10.4.0  # Image handling
